+++ RELEASE_NOTES.md
@@ -32,6 +32,8 @@ Besides the new things listed below, there are a bunch of new extensions, images
 * `KeyboardInputViewController` has a new `keyboardLayoutProvider`.
 * `KeyboardInputViewController` has a new `keyboardSecondaryInputActionProvider`.
 * `KeyboardInputViewController` has a new `keyboardSecondaryInputCalloutContext`.
+* `KeyboardInputViewController` has new properties for `keyboardLocale` and `keyboardType`.
+* `KeyboardInputViewController` has new functions for `changeKeyboardLocale(to:)` and `changeKeyboardType(to:)`.
 * `Layout` has new types for working with keyboard layouts.
 * `Locale` is a new namespace with a few new locale-specific utils.
 * `LocaleDictionary` is a simple dictionary wrapper to work with localized resources.